wip improved worker sheet
